+++ serpentTools/objects/detectors.py
@@ -19,16 +19,24 @@ Detectors
 * :class:`~serpentTools.objects.detectors.SphericalDetector`
 """
 
+from math import sqrt, pi
 from warnings import warn
 from collections import OrderedDict
 
 from six import iteritems
-from numpy import unique, array
+from numpy import unique, array, empty, inf
+from matplotlib.figure import Figure
+from matplotlib.patches import RegularPolygon
+from matplotlib.collections import PatchCollection
+from matplotlib.pyplot import axes
 
 from serpentTools.utils import linkToWiki
 from serpentTools.messages import warning, debug, SerpentToolsException
 from serpentTools.objects.base import DetectorBase
-
+from serpentTools.plot import (
+    magicPlotDocDecorator, formatPlot, setAx_xlims, setAx_ylims,
+    addColorbar, normalizerFactory,
+    )
 
 __all__ = ['Detector', 'CartesianDetector', 'HexagonalDetector',
            'CylindricalDetector', 'SphericalDetector']
@@ -160,9 +168,9 @@ class HexagonalDetector(Detector):
         Type of hexagonal mesh.
 
             2. Flat face perpendicular to x-axis
-            3. Flat face perpendicular to y-axis"""
+            3. Flat face perpendicular to y-axis."""
     __doc__ = """
-    Class that stores detecto data containing a hexagonal meshing.
+    Class that stores detector data containing a hexagonal meshing.
 
     .. versionadded:: 0.5.1
 
@@ -183,37 +191,173 @@ class HexagonalDetector(Detector):
                           detAttrs=Detector.docAttrs,
                           baseAttrs=DetectorBase.baseAttrs)
 
-    def __init__(self, name):
-        Detector.__init__(self, name)
-        self.pitch = None
-        self.hexType = None
-
     _INDEX_MAP = {
         8: 'ycoord',
         9: 'xcoord',
     }
+
     _NON_CART = _INDEX_MAP.values()
+
+    def __init__(self, name):
+        Detector.__init__(self, name)
+        self.__pitch = None
+        self.__hexType = None
+        self.__hexRot = None
+
+    @property
+    def pitch(self):
+        return self.__pitch
+
+    @pitch.setter
+    def pitch(self, value):
+        f = float(value)
+        if f <= 0:
+            raise ValueError("Pitch must be positive")
+        self.__pitch = f
+
+    @property
+    def hexType(self):
+        return self.__hexType
+
+    @hexType.setter
+    def hexType(self, value):
+        if value not in {2, 3}:
+            raise ValueError("Hex type must be 2 or 3, not {}"
+                             .format(value))
+        self.__hexType = value
+        self.__hexRot = 0 if value == 2 else (pi / 2)
 
     def meshPlot(self, xdim, ydim, what='tallies', fixed=None, ax=None,
                  cmap=None, logColor=False, xlabel=None, ylabel=None,
                  logx=False, logy=False, loglog=False, title=None, **kwargs):
+        opts = dict(what=what,
+                    fixed=fixed,
+                    ax=ax,
+                    cmap=cmap,
+                    logColor=logColor,
+                    xlabel=xlabel,
+                    logx=logx,
+                    logy=logy,
+                    loglog=loglog,
+                    title=title,
+                    **kwargs)
         if xdim in self._NON_CART or ydim in self._NON_CART:
-            warnMeshPlot('Hexagonal', 168)
-        return DetectorBase.meshPlot(self,
-                                     xdim, ydim,
-                                     what=what,
-                                     fixed=fixed,
-                                     ax=ax,
-                                     cmap=cmap,
-                                     logColor=logColor,
-                                     xlabel=xlabel,
-                                     logx=logx,
-                                     logy=logy,
-                                     loglog=loglog,
-                                     title=title,
-                                     **kwargs)
+            return self.hexPlot(**opts)
+        return DetectorBase.meshPlot(self, xdim, ydim, **opts)
 
     meshPlot.__doc__ = DetectorBase.meshPlot.__doc__
+
+    @magicPlotDocDecorator
+    def hexPlot(self, what='tallies', fixed=None, ax=None, cmap=None,
+                logColor=False, xlabel=None, ylabel=None, logx=False,
+                logy=False, loglog=False, title=None, normalizer=None,
+                cbarLabel=None, borderpad=2.5, **kwargs):
+        """
+        Create and return a hexagonal mesh plot.
+
+        Parameters
+        ----------
+        what: {'tallies', 'errors', 'scores'}
+            Quantity to plot
+        fixed: None or dict
+            Dictionary of slicing arguments to pass to :meth:`slice`
+        {ax}
+        {cmap}
+        {logColor}
+        {xlabel}
+        {ylabel}
+        {logx}
+        {logy}
+        {loglog}
+        {title}
+        borderpad: int or float
+            Percentage of total plot to apply as a border. A value of
+            zero means that the extreme edges of the hexagons will touch
+            the x and y axis.
+        {kwargs} :class:`matplotlib.patches.RegularPolygon`
+
+        Raises
+        ------
+        AttributeError
+            If :attr:`pitch` and :attr:`hexType` are not set.
+        """
+        borderpad = max(0, float(borderpad))
+        if fixed and ('xcoord' in fixed or 'ycoord' in fixed):
+            raise KeyError("Refusing to restrict along one of the hexagonal "
+                           "dimensions {x/y}coord")
+        for attr in {'pitch', 'hexType'}:
+            if getattr(self, attr) is None:
+                raise AttributeError("{} is not set.".format(attr))
+
+        for key in {'color', 'fc', 'facecolor', 'orientation'}:
+            checkClearKwargs(key, 'hexPlot', **kwargs)
+        ec = kwargs.get('ec', None) or kwargs.get('edgecolor', None)
+        if ec is None:
+            ec = 'k'
+        kwargs['ec'] = kwargs['edgecolor'] = ec
+        if 'figure' in kwargs and kwargs['figure'] is not None:
+            fig = kwargs['figure']
+            if not isinstance(fig, Figure):
+                raise TypeError(
+                        "Expected 'figure' to be of type Figure, is {}"
+                        .format(type(fig)))
+            if len(fig.axes) != 1 and not ax:
+                raise TypeError("Don't know where to place the figure since"
+                                "'figure' argument has multiple axes.")
+            if ax and fig.axes and ax not in fig.axes:
+                raise IndexError("Passed argument for 'figure' and 'ax', "
+                                 "but ax is not attached to figure.")
+            ax = ax or (fig.axes[0] if fig.axes else axes())
+        alpha = kwargs.get('alpha', None)
+
+        ny = len(self.indexes['ycoord'])
+        nx = len(self.indexes['xcoord'])
+        data = self.slice(fixed, what)
+        if data.shape != (ny, nx):
+            raise IndexError("Constrained data does not agree with hexagonal "
+                             "grid structure. Coordinate grid: {}. "
+                             "Constrained shape: {}"
+                             .format((ny, nx), data.shape))
+        nItems = ny * nx
+        patches = empty(nItems, dtype=object)
+        values = empty(nItems)
+        coords = self.grids['COORD']
+
+        ax = ax or axes()
+        pos = 0
+        xmax, ymax = [-inf, ] * 2
+        xmin, ymin = [inf, ] * 2
+        radius = self.pitch / sqrt(3)
+
+        for xy, val in zip(coords, data.flat):
+            values[pos] = val
+            h = RegularPolygon(xy, 6, radius, self.__hexRot, **kwargs)
+            verts = h.get_verts()
+            vmins = verts.min(0)
+            vmaxs = verts.max(0)
+            xmax = max(xmax, vmaxs[0])
+            xmin = min(xmin, vmins[0])
+            ymax = max(ymax, vmaxs[1])
+            ymin = min(ymin, vmins[1])
+            patches[pos] = h
+            pos += 1
+        normalizer = normalizerFactory(values, normalizer, logColor,
+                                       coords[:, 0], coords[:, 1])
+        pc = PatchCollection(patches, cmap=cmap, alpha=alpha)
+        pc.set_array(values)
+        pc.set_norm(normalizer)
+        ax.add_collection(pc)
+
+        addColorbar(ax, pc, None, cbarLabel)
+
+        formatPlot(ax, loglog=loglog, logx=logx, logy=logy,
+                   xlabel=xlabel or "X [cm]",
+                   ylabel=ylabel or "Y [cm]", title=title,
+                   )
+        setAx_xlims(ax, xmin, xmax, pad=borderpad)
+        setAx_ylims(ax, ymin, ymax, pad=borderpad)
+
+        return ax
 
 
 class CylindricalDetector(Detector):
@@ -371,3 +515,10 @@ def warnMeshPlot(plotType, ghIssue):
     msg = ("{} plotting is not fully supported yet - #{}"
            .format(plotType, ghIssue))
     warn(msg, FutureWarning)
+
+
+def checkClearKwargs(key, fName, **kwargs):
+    """Log a warning and clear non-None values from kwargs"""
+    if key in kwargs and kwargs[key] is not None:
+        warning("{} will be set by {}. Worry not.".format(key, fName))
+        kwargs[key] = None