+++ serpentTools/tests/test_ResultsReader.py
@@ -8,7 +8,7 @@ import numpy
 import six
 
 from serpentTools.settings import rc
-from serpentTools.tests import TEST_ROOT
+from serpentTools.data import getFile
 from serpentTools.parsers import ResultsReader
 from serpentTools.messages import SerpentToolsException
 
@@ -26,7 +26,7 @@ class TestBadFiles(unittest.TestCase):
 
     def test_noResults(self):
         """Verify that the reader raises error when no results exist in the file"""
-        badFile = os.path.join(TEST_ROOT, 'bad_results_file.m')
+        badFile = 'bad_results_file.m'
         with open(badFile, 'w') as badObj:
             for _line in range(5):
                 badObj.write(str(_line))
@@ -37,7 +37,7 @@ class TestBadFiles(unittest.TestCase):
 
     def test_noUniverses(self):
         """Verify that the reader raises an error if no universes are stored on the file"""
-        univFile = os.path.join(TEST_ROOT, 'pwr_res_noUniv.m')
+        univFile = getFile('pwr_noUniv_res.m')
         univReader = ResultsReader(univFile)
         with self.assertRaises(SerpentToolsException):
             univReader.read()
@@ -53,7 +53,7 @@ class TestEmptyAttributes(unittest.TestCase):
 
     def test_emptyAttributes(self):
         """Verify that the reader raises error when all attributes are empty"""
-        testFile = os.path.join(TEST_ROOT, 'pwr_res_emptyAttributes.m')
+        testFile = getFile('pwr_emptyAttributes_res.m')
         with self.assertRaises(SerpentToolsException):
             with rc:
                 rc['xs.variableExtras'] = ['GC_UNIVERSE_NAME']
@@ -77,7 +77,7 @@ class TestGetUniv(unittest.TestCase):
                     no universe state exist in the reader
     """
     def setUp(self):
-        self.file = os.path.join(TEST_ROOT, 'pwr_res.m')
+        self.file = getFile('pwr_res.m')
         with rc:
             rc['serpentVersion'] = '2.1.29'
             rc['xs.variableGroups'] = ['versions', 'gc-meta', 'xs',
@@ -206,7 +206,7 @@ class TestFilterResults(TesterCommonResultsReader):
                         univ is filtered
     """
     def setUp(self):
-        self.file = os.path.join(TEST_ROOT, 'pwr_res.m')
+        self.file = getFile('pwr_res.m')
         # universe id, burnup, step, days
         self.expectedStates = (('0', 0.0, 1, 0.0), ('0', 500, 2, 5.0))
         with rc:
@@ -237,14 +237,14 @@ class TestFilterResults(TesterCommonResultsReader):
 
         self.expectedMetadata = {'version': 'Serpent 2.1.29',
                             'compileDate': 'Jan  4 2018 17:22:46',
-                            'debug': [0.],
+                            'debug': 0,
                             'title': 'pwr pin',
-                            'confidentialData': [0.],
+                            'confidentialData': 0,
                             'inputFileName': 'pwrPin',
                             'workingDirectory': '/home/ajohnson400/research/gpt-dep/testing/depmtx',
                             'hostname': 'ME04L0358GRD04',
                             'cpuType': 'Intel(R) Core(TM) i7-6700T CPU @ 2.80GHz',
-                            'cpuMhz': [194.],
+                            'cpuMhz': 194.,
                             'startDate': 'Mon Feb 19 15:39:23 2018',
                             'completeDate': 'Mon Feb 19 15:39:53 2018'}
 
@@ -301,7 +301,7 @@ class TestReadAllResults(TesterCommonResultsReader):
                         univ is not filtered
     """
     def setUp(self):
-        self.file = os.path.join(TEST_ROOT, 'pwr_res_filter.m')
+        self.file = getFile('pwr_filter_res.m')
         # universe id, burnup, step, days
         with rc:
             rc['serpentVersion'] = '2.1.29'
@@ -313,14 +313,14 @@ class TestReadAllResults(TesterCommonResultsReader):
 
         self.expectedMetadata = {'version': 'Serpent 2.1.29',
                             'compileDate': 'Jan  4 2018 17:22:46',
-                            'debug': [0.],
+                            'debug': 0,
                             'title': 'pwr pin',
-                            'confidentialData': [0.],
+                            'confidentialData': 0,
                             'inputFileName': 'pwrPin',
                             'workingDirectory': '/home/ajohnson400/research/gpt-dep/testing/depmtx',
                             'hostname': 'ME04L0358GRD04',
                             'cpuType': 'Intel(R) Core(TM) i7-6700T CPU @ 2.80GHz',
-                            'cpuMhz': [194.],
+                            'cpuMhz': 194.,
                             'startDate': 'Mon Feb 19 15:39:23 2018',
                             'completeDate': 'Mon Feb 19 15:39:53 2018'}
 
@@ -371,7 +371,7 @@ class TestFilterResultsNoBurnup(TesterCommonResultsReader):
                         univ is filtered
     """
     def setUp(self):
-        self.file = os.path.join(TEST_ROOT, 'pwr_res_noBU.m')
+        self.file = getFile('pwr_noBU_res.m')
         # universe id, Idx, Idx, Idx
         self.expectedStates = (('0', 1, 1, 1), ('0', 1, 1, 1))
         with rc:
@@ -402,14 +402,14 @@ class TestFilterResultsNoBurnup(TesterCommonResultsReader):
 
         self.expectedMetadata = {'version': 'Serpent 2.1.30',
                             'compileDate': 'Apr  4 2018 08:55:27',
-                            'debug': [0.],
+                            'debug': 0,
                             'title': 'UO2 PIN MODEL',
-                            'confidentialData': [0.],
+                            'confidentialData': 0,
                             'inputFileName': 'pwr',
                             'workingDirectory': '/gpfs/pace1/project/me-kotlyar/dkotlyar6/Research/Serpent_test/FP_test',
                             'hostname': 'rich133-c36-10-l.pace.gatech.edu',
                             'cpuType': 'Intel(R) Xeon(R) CPU E5-2680 v4 @ 2.40GHz',
-                            'cpuMhz': [184549409.0],
+                            'cpuMhz': 184549409.0,
                             'startDate': 'Mon May 14 11:20:06 2018',
                             'completeDate': 'Mon May 14 11:20:36 2018'}